Script stuff + movement/interaction
Added Day 1's dialogue to a script file

Move most of the player movement code to unhandled_input: move_and_slide only seems to work in process. But now the player won't stop sliding everywhere.

On one hand, the interactions dialog does stop the player from moving as long as the dialog is up... but because the item doesn't go away, the player is stuck looping the dialog over and over.

Maybe try disabling the player's interaction box as soon as the dialog pops up?