Added dialogue scripts for Day 5 and the ending
Gigi's name doesn't show up on the right character label in the ending, or any scene where she talks.

As it stands, the left character label checks for Demi to be in the character position in the dialogue line. I don't know how to check for another character in the same dialogue