Added some audio, main menu
Added sound effect for picking up item (on player scene), talking blips for dialogue (will not play if character name is empty), ambient background audio for basement and entranceway.

Also have some audio for entering the doors, electric zaps for the intro for night 4 and a futuristic teleporter noise for the teleporter in night 5 (Brandon will have to implement these where he sees fit)

Started setting up the main menu. We might need to change the load button into level select? Or have it hidden initially with a flag